Add `utils` module and move `cumulative_percentage_change` inside
