+++ crates/fuel-gas-price-algorithm/src/v1.rs
@@ -4,7 +4,7 @@ use std::{
     ops::Div,
 };
 
-use crate::cumulative_percentage_change;
+use crate::utils::cumulative_percentage_change;
 
 #[cfg(test)]
 mod tests;
